# src/main/resources/openapi.yaml
# cSpell:disable
openapi: 3.0.3
info:
  title: BigoBooks Ventas API
  description: >-
    Microservicio de ventas de libros. Los schemas de respuesta (OrderDto,
    OrderDetailDto, CouponDto) se reutilizan del common (Entrega 1);
    PromotionDto, PageResponseDto y los DTOs de request son propios de este
    servicio.
  version: 1.0.0
paths:
  /orders:
    post:
      tags: [orders]
      operationId: createOrder
      summary: Crea una venta (pedido) con sus items y cupon opcional
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
      responses:
        '201':
          description: Venta creada en estado PENDING
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags: [orders]
      operationId: listOrders
      summary: Lista ventas paginadas, con filtros opcionales
      parameters:
        - name: status
          in: query
          required: false
          description: Estado de la venta
          schema:
            type: string
            enum: [PENDING, PAID, REFUNDED, CANCELLED]
        - name: warehouseId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: couponCode
          in: query
          required: false
          description: Busqueda por codigo de cupon aplicado (contiene)
          schema:
            type: string
        - name: bookId
          in: query
          required: false
          description: true si la venta contiene al libro indicado
          schema:
            type: integer
            format: int64
        - name: createdFrom
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: createdTo
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
        - name: sort
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Pagina de ventas (content = OrderDto[])
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseDto'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /orders/{id}:
    get:
      tags: [orders]
      operationId: getOrderById
      summary: Obtiene una venta por id con sus detalles
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Venta encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [orders]
      operationId: updateOrder
      summary: Actualiza datos de una venta (solo mientras esta PENDING)
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdateRequest'
      responses:
        '200':
          description: Venta actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [orders]
      operationId: deleteOrder
      summary: Elimina logicamente una venta por id
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '204':
          description: Eliminada
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /orders/{id}/status:
    put:
      tags: [orders]
      operationId: changeOrderStatus
      summary: Cambia el estado de una venta (transiciones validas)
      description: >-
        Transiciones permitidas: PENDING->PAID, PENDING->CANCELLED,
        PAID->REFUNDED. Al confirmar el pago (PAID) se descuenta el stock;
        al reembolsar (REFUNDED) se devuelve.
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderStatusChange'
      responses:
        '200':
          description: Estado actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /orders/{id}/coupons:
    post:
      tags: [orders]
      operationId: applyCouponToOrder
      summary: Aplica un cupon a una venta (solo mientras esta PENDING)
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCouponApply'
      responses:
        '200':
          description: Cupon aplicado; se recalcularon subtotal/descuento/total
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [orders]
      operationId: removeCouponFromOrder
      summary: Quita el cupon aplicado a una venta (solo mientras esta PENDING)
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '204':
          description: Cupon quitado y totales recalculados
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /coupons:
    post:
      tags: [coupons]
      operationId: createCoupon
      summary: Crea un cupon de descuento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CouponCreateRequest'
      responses:
        '201':
          description: Cupon creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CouponDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags: [coupons]
      operationId: listCoupons
      summary: Lista cupones paginados, con filtros opcionales
      parameters:
        - name: code
          in: query
          required: false
          description: Busqueda por codigo (contiene)
          schema:
            type: string
        - name: active
          in: query
          required: false
          description: true = cupones sin vencer (validUntil nulo o futuro)
          schema:
            type: boolean
        - name: minDiscountPercent
          in: query
          required: false
          schema:
            type: integer
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
        - name: sort
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Pagina de cupones (content = CouponDto[])
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseDto'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /coupons/{id}:
    get:
      tags: [coupons]
      operationId: getCouponById
      summary: Obtiene un cupon por id
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Cupon encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CouponDto'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [coupons]
      operationId: updateCoupon
      summary: Actualiza un cupon por id
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CouponUpdateRequest'
      responses:
        '200':
          description: Cupon actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CouponDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [coupons]
      operationId: deleteCoupon
      summary: Elimina logicamente un cupon por id
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '204':
          description: Eliminado
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /promotions:
    post:
      tags: [promotions]
      operationId: createPromotion
      summary: Crea una promocion y le asocia libros
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromotionCreateRequest'
      responses:
        '201':
          description: Promocion creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromotionDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags: [promotions]
      operationId: listPromotions
      summary: Lista promociones paginadas, con filtros opcionales
      parameters:
        - name: name
          in: query
          required: false
          description: Busqueda por nombre (contiene)
          schema:
            type: string
        - name: active
          in: query
          required: false
          description: true = promociones vigentes hoy (startDate <= hoy <= endDate)
          schema:
            type: boolean
        - name: minDiscountPercent
          in: query
          required: false
          schema:
            type: integer
        - name: bookId
          in: query
          required: false
          description: true si la promocion incluye al libro indicado
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
        - name: sort
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Pagina de promociones (content = PromotionDto[])
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseDto'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /promotions/{id}:
    get:
      tags: [promotions]
      operationId: getPromotionById
      summary: Obtiene una promocion por id con sus libros
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Promocion encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromotionDto'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [promotions]
      operationId: updatePromotion
      summary: Actualiza una promocion por id
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromotionUpdateRequest'
      responses:
        '200':
          description: Promocion actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromotionDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [promotions]
      operationId: deletePromotion
      summary: Elimina logicamente una promocion por id
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '204':
          description: Eliminada
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /promotions/{id}/books:
    post:
      tags: [promotions]
      operationId: addBooksToPromotion
      summary: Asocia libros a una promocion
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromotionBookAdd'
      responses:
        '201':
          description: Libros asociados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromotionDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /promotions/{id}/books/{bookId}:
    delete:
      tags: [promotions]
      operationId: removeBookFromPromotion
      summary: Desasocia un libro de una promocion
      parameters:
        - $ref: '#/components/parameters/PathId'
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Libro desasociado
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
  responses:
    BadRequest:
      description: Peticion invalida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorDto'
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorDto'
    Conflict:
      description: Operacion no permitida en el estado actual del recurso
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorDto'
    InternalServerError:
      description: Error interno
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorDto'
  schemas:
    ApiErrorDto:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        path:
          type: string
        fieldErrors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    PageResponseDto:
      type: object
      description: Envoltura de paginacion (content es una lista del DTO correspondiente, segun el endpoint)
      properties:
        content:
          type: array
          items:
            type: object
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    OrderCreateRequest:
      type: object
      properties:
        warehouseId:
          type: integer
          format: int64
          description: Almacen de salida (opcional)
        userId:
          type: integer
          format: int64
          description: Usuario que realiza la compra (opcional)
        couponCode:
          type: string
          description: Codigo de cupon opcional para aplicar al total
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemCreate'
      required:
        - items
    OrderItemCreate:
      type: object
      properties:
        bookId:
          type: integer
          format: int64
        quantity:
          type: integer
          format: int32
          description: Cantidad a comprar (mayor a 0)
      required:
        - bookId
        - quantity
    OrderUpdateRequest:
      type: object
      properties:
        warehouseId:
          type: integer
          format: int64
          description: Nuevo almacen de salida (null = sin cambios)
        userId:
          type: integer
          format: int64
          description: Nuevo usuario de la venta (null = sin usuario)
        bancardNumber:
          type: string
        invoiceNumber:
          type: string
    OrderStatusChange:
      type: object
      properties:
        status:
          type: string
          enum: [PENDING, PAID, REFUNDED, CANCELLED]
      required:
        - status
    OrderCouponApply:
      type: object
      properties:
        couponCode:
          type: string
      required:
        - couponCode
    CouponCreateRequest:
      type: object
      properties:
        code:
          type: string
          description: Codigo unico del cupon
        discountPercent:
          type: integer
          description: Descuento entre 1 y 100
        validUntil:
          type: string
          format: date-time
          description: Vencimiento (opcional; nulo = sin vencimiento)
      required:
        - code
        - discountPercent
    CouponUpdateRequest:
      type: object
      properties:
        discountPercent:
          type: integer
        validUntil:
          type: string
          format: date-time
          description: Vencimiento (null = sin vencimiento)
    PromotionCreateRequest:
      type: object
      properties:
        name:
          type: string
        discountPercent:
          type: integer
          description: Descuento entre 1 y 100
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        bookIds:
          type: array
          description: Libros que participan de la promocion
          items:
            type: integer
            format: int64
      required:
        - name
        - discountPercent
        - startDate
        - endDate
    PromotionUpdateRequest:
      type: object
      properties:
        name:
          type: string
        discountPercent:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
    PromotionBookAdd:
      type: object
      properties:
        bookIds:
          type: array
          items:
            type: integer
            format: int64
      required:
        - bookIds
    PromotionDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        discountPercent:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
        bookIds:
          type: array
          items:
            type: integer
            format: int64
    # Los siguientes schemas se documentan aca pero NO se generan
    # (modelsToGenerate): sus clases viven en el JAR del common.
    OrderDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        warehouseId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        status:
          type: string
          enum: [PENDING, PAID, REFUNDED, CANCELLED]
        totalPrice:
          type: integer
          format: int64
        subtotal:
          type: integer
          format: int64
        discountAmount:
          type: integer
          format: int64
        bancardNumber:
          type: string
        invoiceNumber:
          type: string
        createdAt:
          type: string
          format: date-time
        couponId:
          type: integer
          format: int64
        orderDetails:
          type: array
          items:
            $ref: '#/components/schemas/OrderDetailDto'
    OrderDetailDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        bookId:
          type: integer
          format: int64
        bookName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: integer
          format: int64
        discount:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        orderId:
          type: integer
          format: int64
    CouponDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        discountPercent:
          type: integer
        validUntil:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time